--- src/main/java/org/example/mjuteam4/myPlant/MyPlantController.java
@@ -1,4 +1,0 @@
-package org.example.mjuteam4.myPlant;
-
-public class MyPlantController {
-}
