+++ src/main/java/org/example/mjuteam4/mypage/entity/Member.java
@@ -1,10 +1,9 @@
-package org.example.mjuteam4.member;
+package org.example.mjuteam4.mypage.entity;
 
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.mjuteam4.myPlant.MyPlant;
 
 import java.util.ArrayList;
 import java.util.List;
