+++ src/main/java/org/example/mjuteam4/mypage/exception/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package org.example.mjuteam4.member.exception;
+package org.example.mjuteam4.mypage.exception;
 
 import org.example.mjuteam4.global.exception.ExceptionCode;
 import org.example.mjuteam4.global.exception.GlobalException;
