+++ src/main/java/org/example/mjuteam4/mypage/repository/MyPlantRepository.java
@@ -1,5 +1,6 @@
-package org.example.mjuteam4.myPlant;
+package org.example.mjuteam4.mypage.repository;
 
+import org.example.mjuteam4.mypage.entity.MyPlant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MyPlantRepository extends JpaRepository<MyPlant, Long> {
