+++ src/main/java/org/example/mjuteam4/security/handler/OAuth2SuccessHandler.java
@@ -6,8 +6,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.mjuteam4.global.uitl.JwtUtil;
-import org.example.mjuteam4.member.Member;
-import org.example.mjuteam4.member.MemberRepository;
+import org.example.mjuteam4.mypage.entity.Member;
+import org.example.mjuteam4.mypage.repository.MemberRepository;
 import org.example.mjuteam4.security.provider.TokenProvider;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
