+++ src/main/java/org/example/mjuteam4/security/provider/TokenProvider.java
@@ -9,10 +9,9 @@ import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.mjuteam4.global.uitl.JwtUtil;
-import org.example.mjuteam4.member.Member;
-import org.example.mjuteam4.member.MemberRepository;
-import org.example.mjuteam4.member.exception.MemberNotFoundException;
-import org.example.mjuteam4.redis.Token;
+import org.example.mjuteam4.mypage.entity.Member;
+import org.example.mjuteam4.mypage.repository.MemberRepository;
+import org.example.mjuteam4.mypage.exception.MemberNotFoundException;
 import org.example.mjuteam4.redis.TokenRepository;
 import org.example.mjuteam4.redis.TokenService;
 import org.example.mjuteam4.redis.exception.JwtSignatureInvalidException;
